requirements.txt
alembic==0.8.10
apache-airflow==1.8.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
croniter==0.3.17
dill==0.2.7.1
docutils==0.13.1
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.12
funcsigs==1.0.0
future==0.15.2
gitdb2==2.0.2
GitPython==2.1.5
gunicorn==19.3.0
idna==2.5
itsdangerous==0.24
Jinja2==2.8.1
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
numpy==1.13.1
ordereddict==1.1
pandas==0.20.3
pkg-resources==0.0.0
psutil==4.4.2
psycopg2==2.7.3
Pygments==2.2.0
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyYAML==3.12
requests==2.18.1
setproctitle==1.1.10
six==1.10.0
smmap2==2.0.3
SQLAlchemy==1.1.11
tabulate==0.7.7
thrift==0.9.3
Unidecode==0.4.21
urllib3==1.21.1
Werkzeug==0.12.2
WTForms==2.1
zope.deprecation==4.2.0
